+++ plugins/index/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
 
+## Unreleased
+ - Show PID, event loop lag, resident set size and heap usage.
+ - Allow configuration of welcome message.
+
 ## 1.0.0
  - Initial release